Changes to generate the JuliaSet

+++ Fractalis.Domain/Base/BaseComplex.cs
@@ -4,7 +4,7 @@ namespace Fractalis.Domain.Base
 {
     public class BaseComplex : IComplex
     {
-        public int Real { get; set; }
-        public int Imaginary { get; set; }
+        public double Real { get; set; }
+        public double Imaginary { get; set; }
     }
 }
